mesh_seq: use options prefixes
