+++ pyroteus/time_partition.py
@@ -33,10 +33,6 @@ class TimePartition(object):
             timesteps per export (default 1)
         :kwarg start_time: start time of the
             interval of interest (default 0.0)
-        :kwarg solves_per_timestep: (list of)
-            (non)linear solves per timestep
-            corresponding to the fields (defaults
-            to 1 for each)
         :kwarg subinterals: user-provided sequence
             of subintervals, which need not be of
             uniform length (defaults to None)
@@ -50,13 +46,6 @@ class TimePartition(object):
         self.num_subintervals = int(np.round(num_subintervals))
         if not np.isclose(num_subintervals, self.num_subintervals):
             raise ValueError(f"Non-integer number of subintervals {num_subintervals}")
-        solves_per_timestep = kwargs.get('solves_per_timestep', [1 for field in fields])
-        if not isinstance(solves_per_timestep, Iterable):
-            solves_per_timestep = [solves_per_timestep]
-        self.solves_per_timestep = [int(np.round(spts)) for spts in solves_per_timestep]
-        if not np.allclose(solves_per_timestep, self.solves_per_timestep):
-            raise ValueError(f"Non-integer number of solves per timestep {solves_per_timestep}")
-        self.print("solves_per_timestep")
         self.debug("num_subintervals")
         self.interval = (self.start_time, self.end_time)
         self.debug("interval")
@@ -156,61 +145,6 @@ class TimePartition(object):
             'start_time': self.subintervals[i][0],
             'end_time': self.subintervals[i][1],
         })
-
-    def get_solve_blocks(self, field, subinterval=0, has_adj_sol=True):
-        """
-        Get all blocks of the tape corresponding to
-        solve steps for prognostic solution ``field``
-        on a given ``subinterval``.
-        """
-        from firedrake.adjoint.blocks import GenericSolveBlock, ProjectBlock
-        from pyadjoint import get_working_tape
-
-        # Get all blocks
-        blocks = get_working_tape().get_blocks()
-        if len(blocks) == 0:
-            pyrint("WARNING: tape has no blocks!")
-            return blocks
-
-        # Restrict to solve blocks
-        solve_blocks = [
-            block
-            for block in blocks
-            if issubclass(block.__class__, GenericSolveBlock)
-            and not issubclass(block.__class__, ProjectBlock)
-        ]
-
-        # Restrict to solve blocks with adjoint solutions
-        if has_adj_sol:
-            solve_blocks = [
-                block
-                for block in solve_blocks
-                if block.adj_sol is not None
-            ]
-
-        # Slice solve blocks by field
-        stride = sum(self.solves_per_timestep)
-        offset = sum(self.solves_per_timestep[:self.fields.index(field) + 1])
-        offset -= self.timesteps_per_subinterval[subinterval]*stride
-        if self.debug:
-            pyrint("Solve blocks before slicing:")
-            for i, block in enumerate(solve_blocks):
-                pyrint(f"{i:4d}: {type(block)} {block.options_prefix}")
-            pyrint(f"Offset = {offset}")
-            pyrint(f"Stride = {stride}")
-        solve_blocks = solve_blocks[offset::stride]
-        if self.debug:
-            pyrint("Solve blocks after slicing:")
-            for i, block in enumerate(solve_blocks):
-                pyrint(f"{i:4d}: {type(block)} {block.options_prefix}")
-
-        # Check FunctionSpaces are consistent across solve blocks
-        element = solve_blocks[0].function_space.ufl_element()
-        for block in solve_blocks:
-            if element != block.function_space.ufl_element():
-                raise ValueError(f"Solve block list for field {field} contains mismatching"
-                                 + f" elements ({element} vs. {block.function_space.ufl_element()})")
-        return solve_blocks
 
 
 class TimeInterval(TimePartition):
